+++ inst/examples/example-indexSelfConstruction.R
@@ -7,6 +7,9 @@ indexSelfConstruction(boeker, 1, 2, c(3:11), method = "minkowski", p = 3)
 indexSelfConstruction(boeker, 1, 2, c(3:11), method = "pearson")
 indexSelfConstruction(boeker, 1, 2, c(3:11), method = "spearman")
 
+# using not-normalized distances
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "euclidean", normalize = FALSE)
+
 # printing the results (biplot only works with)
 cp <- indexSelfConstruction(boeker, 1, 2, c(3:11))
 cp$grid  # grid with self, ideal and others